+++ src/imio/smartweb/core/contents/pages/page/content.py
@@ -6,13 +6,15 @@ from imio.smartweb.locales import SmartwebMessageFactory as _
 from zope import schema
 from zope.interface import implementer
 
+CATEGORY_TAXONOMY = "collective.taxonomy.page"
+
 
 class IPage(IPages):
     """Marker interface and Dexterity Python Schema for Page"""
 
     category = schema.Choice(
         title=_(u"Category"),
-        source="collective.taxonomy.page",
+        source=CATEGORY_TAXONOMY,
         required=False,
     )
 
@@ -20,3 +22,4 @@ class IPage(IPages):
 @implementer(IPage)
 class Page(Pages):
     """Page class"""
+    category_taxonomy = CATEGORY_TAXONOMY